feat(display): implement dynamic font scaling based on available space
- Change `FontDetails.size` from `int` to `double` to allow fine-grained, fractional font sizes.
- Introduce `setFontToFit(width, height)` in `FontManager` interface and implement it in `FontManagerType1` using binary search to find the largest fitting font size.
- Refactor `Screen.fontChanged()` to prevent unnecessary layout jumps when resizing via window dragging, while preserving legacy menu-driven resizing behavior.
- Add `Screen.resizeToFit()` to calculate max character constraints from available space.

+++ src/com/bytezone/dm3270/display/FontDetails.java
@@ -13,9 +13,9 @@ public class FontDetails
   public final Font font;
 
   public final String name;
-  public final int size;
+  public final double size;
 
-  public FontDetails (String name, int size, Font font)
+  public FontDetails (String name, double size, Font font)
   {
     this.font = font;
     this.name = name;
